Refactor code, move code to the base class (#2905)
